pyproject.toml
[tool.poetry]
name = "ovino"
version = "0.1.0"
description = "Projeto de ovinocultura (NEPPA - Fabrica de Software)"
authors = ["Eduardo da Silva <eduardo.silva@ifc.edu.br>"]

[tool.poetry.dependencies]
python = ">=3.7,<3.9"
django = "^3.0.5"
djangorestframework = "^3.11.0"
drf-yasg = "^1.17.1"
djangorestframework-simplejwt = "^4.4.0"
rules = "^2.2"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.7.9"
isort = "^4.3.21"
pre-commit = "^2.2.0"
pytest = "^5.4.1"
pytest-cov = "^2.8.1"
pytest-django = "^3.9.0"

[tool.black]
line-length = 119

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 119
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

